+++ packages/core/src/core/config.js
@@ -2,8 +2,6 @@
 export const config = {
   url: 'http://127.0.0.1:8080/api',
   projectName: '',
-  appId: '123456',
-  userId: '123456',
   isImageUpload: false,
   batchSize: 20
 };