+++ packages/core/src/index.js
@@ -6,9 +6,7 @@ import {
   overWriteXhr,
   overWriteFetch
 } from './core/index.js';
-window.__Tmonitor__={
-
-};
+import {Tmonitor} from '@web-tmonitor/utils';
 
 function report(type) {
   return function (res) {
@@ -30,15 +28,19 @@ export default {
   },
   use(plugin, config){
     const instance = new plugin(config);
-    instance.core({
-      eventBus: this.eventBus,
-      reportData
-    });
+    try{
+      instance.core({
+        eventBus: this.eventBus,
+        reportData
+      });
+    }catch (e) {
+
+    }
   },
   // 针对Vue项目的错误捕获
   install(Vue, options) {
-    if (__Tmonitor__.vue) return;
-    __Tmonitor__.vue = true;
+    if (Tmonitor.vue) return;
+    Tmonitor.vue = true;
     setConfig(options);
     const handler = Vue.config.errorHandler;
     // vue项目中 通过 Vue.config.errorHandler 捕获错误
@@ -60,8 +62,8 @@ export default {
   },
   // 针对React项目的错误捕获
   errorBoundary(err, info) {
-    if (__Tmonitor__.react) return;
-    __Tmonitor__.react = true;
+    if (Tmonitor.react) return;
+    Tmonitor.react = true;
     // 上报具体的错误信息
     const data = {
       error: err?.stack,