+++ packages/utils/src/globalVar.js
@@ -1,0 +1,7 @@
+
+export const Tmonitor = getTmonitor();
+
+export function getTmonitor(){
+  window.__Tmonitor__ = window.__Tmonitor__||{};
+  return window.__Tmonitor__;
+}