+++ packages/core/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @tmonitor/core
 
+## 1.1.1
+
+### Patch Changes
+
+- user behavior and whiteSreen monitoring
+- Updated dependencies
+  - @web-tmonitor/performance@1.1.2
+  - @web-tmonitor/utils@1.0.4
+
+## 1.1.0
+
+### Minor Changes
+
+- Error and interface monitoring
+
+### Patch Changes
+
+- Updated dependencies
+  - @web-tmonitor/performance@1.1.1
+  - @web-tmonitor/utils@1.0.3
+
 ## 1.0.3
 
 ### Patch Changes