+++ packages/record/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @web-tmonitor/record
 
+## 1.0.2
+
+### Patch Changes
+
+- user behavior and whiteSreen monitoring
+
 ## 1.0.1
 
 ### Patch Changes