+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @web-tmonitor/utils
 
+## 1.0.4
+
+### Patch Changes
+
+- user behavior and whiteSreen monitoring
+
+## 1.0.3
+
+### Patch Changes
+
+- Error and interface monitoring
+
 ## 1.0.2
 
 ### Patch Changes